requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
Django==4.2.6
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
pytest==7.4.4
pytest-django==4.7.0
python-dotenv==1.0.0
sqlparse==0.4.4
yarl==1.9.2
